refactor: 콘텐츠 상세 페이지 커스텀 훅 폴더 세분화
